Began the design of the basic game loop. Allows user to flag mines that they have found as well as remove flags. Seperated gamestate into seperate header.

+++ Ex2_Mines/Ex2_Mines2/settingsMenu.cpp
@@ -116,7 +116,7 @@ bool settingsMenu::checkInput(std::string input, int& parameterValue)
 	{
 		if (input.length() >= 3)
 		{
-			for (int i = 1; i < input.length(); i++)
+			for (int i = 1; i < (int)input.length(); i++)
 			{
 				if (whitespaces <= 1)
 				{
